refactor[#47] : 사용하지 않는 features/markdownEditor 컴포넌트 제거
