+++ blog/src/features/article/ui/index.ts
@@ -1,1 +1,1 @@
-export * from "./MarkdownEditor";
+export * from "./ArticleSidebar";
